Signal implementation separated to its own file
