+++ include/signal_manager.hpp
@@ -2,35 +2,17 @@
 #define SIGNAL_MANAGER
 
 #include <vector>
-#include <string>
+#include <map>
 #include <mutex>
+#include "signal.hpp"
+#include "subscriber.hpp"
 
 namespace tbo
 {
-	class signal
-	{
-	private:
-	public:
-		signal()
-		{
-		}
-		signal(const std::string& _recipient, const std::string& _command) : recipient(_recipient), command(_command)
-		{
-		}
-		std::string recipient;
-		std::string command;
-
-		friend bool operator !=(const tbo::signal&, const tbo::signal&);
-		friend bool operator ==(const tbo::signal&, const tbo::signal&);
-		friend std::ostream& operator <<(std::ostream& os, const tbo::signal& sig);
-	};
-	bool operator ==(const tbo::signal& sig1, const tbo::signal& sig2);
-	bool operator !=(const tbo::signal& sig1, const tbo::signal& sig2);
-	std::ostream& operator <<(std::ostream& os, const tbo::signal& sig);
-		
 	class signal_manager
 	{
 	private:
+		static std::map<std::string, std::vector<tbo::subscriber*>> subscriptions; // Storage subscription and subscribers
 		static std::vector<tbo::signal> signals;
 		static std::mutex locker;
 	public:
