+++ source/signal_manager.cpp
@@ -1,7 +1,6 @@
 #include "signal_manager.hpp"
 #include <algorithm>
 #include <mutex>
-#include <ostream>
 
 std::mutex tbo::signal_manager::locker;
 std::vector<tbo::signal> tbo::signal_manager::signals;
@@ -25,20 +24,4 @@ void tbo::signal_manager::add_signal(const tbo::signal& sig)
 	signals.push_back(sig);
 }
 
-bool tbo::operator == (const tbo::signal& sig1, const tbo::signal& sig2)
-{
-	if (sig1.command == sig2.command && sig1.recipient == sig2.recipient) return true;
-	else return false;
-}
 
-bool tbo::operator != (const tbo::signal& sig1, const tbo::signal& sig2)
-{
-	if (!(sig1 == sig2)) return true;
-	else return false;
-}
-
-std::ostream& tbo::operator << (std::ostream& os, const tbo::signal& sig)
-{
-	os << "Recipient: " << sig.recipient << " | " << "Command: " << sig.command;
-	return os;
-}
